Removed original implementations.
The given implementations from the "inspirational" source all required
different parameters, than provided by the closure-library.

As such, the implementations are just "noise" in the code. Also, without
knowing the purpose of the parameters, it's very hard to port them.

Removing them makes it easier to discern what's implemented and what
not.